requirements.txt
asgiref==3.4.1
Django>3.2,<4
pytz==2021.1
sqlparse==0.4.1
Pillow
django-robots
django-debug-toolbar
six
django-taggit==1.4.0
django-tinymce
django-jalali-date